requirements.txt
Flask
requests
python-dotenv
flask_login
psycopg2-binary
Flask-SQLAlchemy
